requirements.txt
black
pytest
pytest-dependency
squareup
requests
